feat: check if raw is registered
if we want to use Async with async_std::net::TcpStream, the Async::new
will return AlreadyExists error because we can't register a fd into the
same twice.
this commit will check if we have registered the fd when register return
AlreadyExists error, if registered, return the clone of the exist Source.